# frontend/openapi/modular-pipeline-openapi.json
{
  "components": {
    "schemas": {
...
      "ContinuationResponse": {
        "properties": {
          "causal_chain_summary": {
            "default": "",
            "title": "Causal Chain Summary",
            "type": "string"
          },
          "evidence_texts": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Evidence Texts",
            "type": "array"
          },
          "related_figures": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Related Figures",
            "type": "array"
          },
          "spark": {
            "$ref": "#/components/schemas/SparkResponse"
          },
          "suggested_angles": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Suggested Angles",
            "type": "array"
          }
        },
        "required": [
          "spark"
        ],
        "title": "ContinuationResponse",
        "type": "object"
      },
...
      "GenerateSparksRequest": {
        "properties": {
          "limit": {
            "default": 10,
            "description": "最大返回数",
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "llm": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LLMConfig"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional LLM config for real inspiration generation"
          },
          "project_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "项目ID，用于从知识库生成启发点",
            "title": "Project Id"
          },
          "query": {
            "description": "查询/主题",
            "maxLength": 500,
            "minLength": 1,
            "title": "Query",
            "type": "string"
          },
          "sampling": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "number"
                    },
                    {
                      "type": "integer"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Per-task sampling overrides",
            "title": "Sampling"
          }
        },
        "required": [
          "query"
        ],
        "title": "GenerateSparksRequest",
        "type": "object"
      },
      "GenerateSparksResponse": {
        "properties": {
          "sparks": {
            "items": {
              "$ref": "#/components/schemas/SparkResponse"
            },
            "title": "Sparks",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "sparks",
          "total"
        ],
        "title": "GenerateSparksResponse",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LLMConfig": {
        "properties": {
          "api_key": {
            "default": "",
            "title": "Api Key",
            "type": "string"
          },
          "base_url": {
            "default": "https://api.deepseek.com",
            "title": "Base Url",
            "type": "string"
          },
          "max_tokens": {
            "default": 4096,
            "title": "Max Tokens",
            "type": "integer"
          },
          "model": {
            "default": "deepseek-chat",
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "default": "DeepSeek",
            "title": "Provider",
            "type": "string"
          },
          "system_prompt": {
            "default": "",
            "title": "System Prompt",
            "type": "string"
          },
          "temperature": {
            "default": 0.7,
            "title": "Temperature",
            "type": "number"
          },
          "top_k": {
            "default": 50,
            "title": "Top K",
            "type": "integer"
          },
          "top_p": {
            "default": 0.9,
            "title": "Top P",
            "type": "number"
          }
        },
        "title": "LLMConfig",
        "type": "object"
      },
...
      "MemorySearchHitPayload": {
        "description": "Single MemPalace search hit.",
        "properties": {
          "room": {
            "title": "Room",
            "type": "string"
          },
          "similarity": {
            "title": "Similarity",
            "type": "number"
          },
          "source_file": {
            "title": "Source File",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          },
          "wing": {
            "title": "Wing",
            "type": "string"
          }
        },
        "required": [
          "text",
          "wing",
          "room",
          "source_file",
          "similarity"
        ],
        "title": "MemorySearchHitPayload",
        "type": "object"
      },
      "MemorySearchRequest": {
        "description": "Semantic memory search request.",
        "properties": {
          "limit": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Limit"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "room": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room"
          },
          "wing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wing"
          }
        },
        "required": [
          "query"
        ],
        "title": "MemorySearchRequest",
        "type": "object"
      },
      "MemorySearchResponsePayload": {
        "description": "Search response payload for MemPalace queries.",
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "query": {
            "title": "Query",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/MemorySearchHitPayload"
            },
            "title": "Results",
            "type": "array"
          },
          "room": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Room"
          },
          "wing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wing"
          }
        },
        "required": [
          "query",
          "available"
        ],
        "title": "MemorySearchResponsePayload",
        "type": "object"
      },
...
      "MemorySyncPayload": {
        "description": "Result payload for runtime-to-memory synchronization.",
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "drawer_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Drawer Id"
          },
          "duplicate": {
            "default": false,
            "title": "Duplicate",
            "type": "boolean"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "room": {
            "title": "Room",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          },
          "wing": {
            "title": "Wing",
            "type": "string"
          }
        },
        "required": [
          "success",
          "available",
          "wing",
          "room"
        ],
        "title": "MemorySyncPayload",
        "type": "object"
      },
      "MemoryWakeupPayload": {
        "description": "Wake-up context payload built from Layer0 + Layer1.",
        "properties": {
          "available": {
            "title": "Available",
            "type": "boolean"
          },
          "context": {
            "title": "Context",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "token_estimate": {
            "title": "Token Estimate",
            "type": "integer"
          },
          "wing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wing"
          }
        },
        "required": [
          "context",
          "token_estimate",
          "available"
        ],
        "title": "MemoryWakeupPayload",
        "type": "object"
      },
...
      "SparkResponse": {
        "properties": {
          "actionable": {
            "default": true,
            "title": "Actionable",
            "type": "boolean"
          },
          "confidence": {
            "default": 0.0,
            "title": "Confidence",
            "type": "number"
          },
          "content": {
            "title": "Content",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "related_point_ids": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Related Point Ids",
            "type": "array"
          },
          "source_papers": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Source Papers",
            "type": "array"
          },
          "spark_type": {
            "title": "Spark Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "content",
          "spark_type"
        ],
        "title": "SparkResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "学术研究智能体 — 论文分析、知识管理与智能写作辅助平台",
    "title": "Scholar AI API",
    "version": "1.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/inspiration/generate": {
      "post": {
        "description": "基于查询生成启发点列表。",
        "operationId": "post_inspiration_generate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateSparksRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateSparksResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Inspirations",
        "tags": [
          "Inspiration"
        ]
      }
    },
...
    "/inspiration/{spark_id}/context": {
      "get": {
        "description": "获取启发点的续写上下文。",
        "operationId": "get_inspiration_spark_id_context",
        "parameters": [
          {
            "in": "path",
            "name": "spark_id",
            "required": true,
            "schema": {
              "title": "Spark Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContinuationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Spark Context",
        "tags": [
          "Inspiration"
        ]
      }
    },
...
    "/memory/runtime/job/{job_id}/sync": {
      "post": {
        "description": "Synchronize evidence from a completed job into long-term memory.",
        "operationId": "post_memory_runtime_job_job_id_sync",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemorySyncPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Sync Runtime Job To Memory",
        "tags": [
          "Memory"
        ]
      }
    },
    "/memory/search": {
      "post": {
        "description": "Search MemPalace for project memory.",
        "operationId": "post_memory_search",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemorySearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemorySearchResponsePayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Memory",
        "tags": [
          "Memory"
        ]
      }
    },
...
    "/memory/wakeup": {
      "get": {
        "description": "Render Layer0 + Layer1 context for the current project wing.",
        "operationId": "get_memory_wakeup",
        "parameters": [
          {
            "in": "query",
            "name": "wing",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Wing"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryWakeupPayload"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Memory Wakeup",
        "tags": [
          "Memory"
        ]
      }
    },
...
  },
...
}